+++ src/screens/ScreenPokedex.tsx
@@ -6,20 +6,21 @@ import { useSessionStorage } from "usehooks-ts";
 import { CopyButton } from "../components/CopyButton";
 import { Divider } from "../components/Divider";
 import { EmptyState } from "../components/EmptyState";
-import { FancyLink } from "../components/FancyLink";
-import { FancyText } from "../components/FancyText";
 import { Flex } from "../components/Flex";
 import { Monster } from "../components/Monster";
 import { Padding } from "../components/Padding";
 import { Paginator } from "../components/Paginator";
 import { Search } from "../components/Search";
+import { Select } from "../components/Select";
 import { Spinner } from "../components/Spinner";
 import { useAppContext } from "../hooks/useAppContext";
 import { useSearch } from "../hooks/useSearch";
+import { compare } from "../misc/compare";
 import { Type, typesFromUserInput } from "../misc/data-types";
 import { formatMonsterNumber } from "../misc/formatMonsterNumber";
 import { pickTranslation } from "../misc/pickTranslation";
 import styles from "./ScreenPokedex.module.css";
+import { FancyLink } from "../components/FancyLink";
 
 export function ScreenPokedex(): ReactNode {
   const { allPokemon, isLoading } = useAppContext();
@@ -27,6 +28,7 @@ export function ScreenPokedex(): ReactNode {
   const { language } = i18n;
   const search = useSearch();
   const [query, setQuery] = useSessionStorage("pokedex.query", "");
+  const [sortOrder, setSortOrder] = useSessionStorage("pokedex.sortOrder", "");
   const deferredQuery = useDeferredValue(query);
   const isStale = query !== deferredQuery;
   const [page, setPage] = useSessionStorage<number>("pokedex.page", 0);
@@ -48,6 +50,7 @@ export function ScreenPokedex(): ReactNode {
         ...p,
         speciesName: pickTranslation(p.speciesNames, language),
         formName: pickTranslation(p.formNames, language),
+        total: p.hp + p.attack + p.defense + p.spAttack + p.spDefense + p.speed,
       };
     });
   }, [allPokemon, language]);
@@ -62,8 +65,9 @@ export function ScreenPokedex(): ReactNode {
     // the rules of hooks can't realize these. Pretend to use `language` here to
     // make it happy.
     const types = typesFromUserInput({ types: s, t, strict: true });
+    let items = searchablePkmn.slice();
     if (types.length > 0) {
-      return searchablePkmn.filter((p) => {
+      items = searchablePkmn.filter((p) => {
         if (types.length === 1) {
           return p.types[0] === types[0] || p.types[1] === types[0];
         }
@@ -74,14 +78,26 @@ export function ScreenPokedex(): ReactNode {
           p.types.slice().sort().join(" ") === types.slice().sort().join(" ")
         );
       });
+    } else if (s) {
+      items = matchSorter(searchablePkmn, s, {
+        keys: ["speciesName", "formName", "number"],
+      });
     }
-    if (!s) {
-      return searchablePkmn;
+    switch (sortOrder) {
+      case "attack":
+      case "defense":
+      case "spAttack":
+      case "spDefense":
+      case "speed":
+      case "total": {
+        items = items.slice().sort((a, b) => {
+          return -compare(a[sortOrder], b[sortOrder]);
+        });
+        break;
+      }
     }
-    return matchSorter(searchablePkmn, s, {
-      keys: ["speciesName", "formName", "number"],
-    });
-  }, [deferredQuery, searchablePkmn, language, t]);
+    return items;
+  }, [deferredQuery, searchablePkmn, language, t, sortOrder]);
 
   const permalink = new URL(window.location.href);
   {
@@ -103,23 +119,31 @@ export function ScreenPokedex(): ReactNode {
     <main className="center content-wide">
       <Flex direction="column" gap="large" padding="large">
         <Padding size="small" />
-        <Flex direction="column" gap="medium">
-          <Search search={query} updateSearch={updateSearch} />
-          <div className={styles.small}>
-            <Flex>
-              <FancyText tag="span" aria-hidden="true">
-                {t("pokedex.search.description")}
-              </FancyText>
-              <Flex flex="auto" />
-              <FancyLink
-                to="/pokedex/help/"
-                aria-label={t("pokedex.search.helpLong")}
-              >
-                {t("pokedex.search.help")}
-              </FancyLink>
-            </Flex>
-          </div>
-        </Flex>
+        <div className={styles.searchArea}>
+          <Flex direction="column" gap="medium">
+            <Search
+              label={t("pokedex.search.description")}
+              value={query}
+              onChange={updateSearch}
+            />
+          </Flex>
+          <Select
+            label={t("pokedex.sorting.sortBy")}
+            value={sortOrder}
+            onChange={(event) => {
+              setSortOrder(event.target.value);
+            }}
+          >
+            <option value="">Auto</option>
+            <hr />
+            <option value="attack">{t("pokedex.stats.attack")}</option>
+            <option value="defense">{t("pokedex.stats.defense")}</option>
+            <option value="spAttack">{t("pokedex.stats.specialAttack")}</option>
+            <option value="spDef">{t("pokedex.stats.specialDefense")}</option>
+            <option value="speed">{t("pokedex.stats.speed")}</option>
+            <option value="total">{t("pokedex.stats.total")}</option>
+          </Select>
+        </div>
         <Divider />
         {isLoading ? (
           <Spinner />
@@ -128,7 +152,9 @@ export function ScreenPokedex(): ReactNode {
             currentPage={Number(page)}
             setPage={setPage}
             pageSize={8 * 3}
-            emptyState={<EmptyState>{t("pokedex.search.notFound")}</EmptyState>}
+            emptyState={
+              <EmptyState borderless>{t("pokedex.search.notFound")}</EmptyState>
+            }
             items={pkmn}
             renderID={(pkmn) => formatMonsterNumber(Number(pkmn.number))}
             renderPage={(page) => (
@@ -144,8 +170,16 @@ export function ScreenPokedex(): ReactNode {
             )}
           />
         )}
-        <Flex>
+        <Divider />
+        <Flex align="center">
           <CopyButton text={permalink.href}>{t("general.copyLink")}</CopyButton>
+          <Flex flex="auto" />
+          <FancyLink
+            to="/pokedex/help/"
+            aria-label={t("pokedex.search.helpLong")}
+          >
+            {t("pokedex.search.help")}
+          </FancyLink>
         </Flex>
       </Flex>
     </main>